GRASS

Documentation for GRASS.

GRASS.AbstractSolarData

GRASS.DiskParams

GRASS.SolarData

GRASS.SpecParams

GRASS.absorption_line

GRASS.calc_ccf

GRASS.calc_intensity_normalization

GRASS.calc_mu

GRASS.calc_r2

GRASS.calc_rms

GRASS.calc_rvs_from_ccf

GRASS.clean_input

GRASS.doppler_factor

GRASS.extrapolate_bisector

GRASS.extrapolate_width

GRASS.find_nearest_ax

GRASS.make_grid

GRASS.mu_map

GRASS.norm_term

GRASS.patch_velocity_los

GRASS.quad_limb_darkening

GRASS.relative_bisector_wavelengths

GRASS.rotation_period

GRASS.searchsortednearest

GRASS.synthesize_spectra

GRASS.uniqueidx

GRASS.wav2vel

GRASS.width_thermal

GRASS.AbstractSolarData — Type

GRASS.DiskParams — Method

DiskParams(; N=256, Nt=50, pole=(0.0, 1.0, 0.0))

Construct a DiskParams composite type instance.

Arguments

N::Integer=256: Length of N*N spatial grid

Nt::Integer=50: Number of 15-second snapshots

pole::Tuple{Float64, Float64, Float64}=(0.0, 1.0, 0.0): Unit vector specificying rotation axis direction. Default is equator-on.

GRASS.SolarData — Method

GRASS.SpecParams — Method

SpecParams(; lines=[], depths=[], variability=[], resolution=7e8)

Construct a SpecParams composite type instance.

Arguments

lines::AbstractArray{Float64,1}=[]: List of line centers (in angstroms)

depths::AbstractArray{Float64,1}=[]: List of line depths

variability::AbstractArray{Bool,1}=[]: Array of booleans controlling whether corresponding line has variability

resolution::Float64=7e8: Spectral resolution of spectrum

GRASS.absorption_line — Method

Absorption line

GRASS.calc_ccf — Method

GRASS.calc_intensity_normalization — Method

Integrate over smooth stellar disk assuming quadratic limb darkening law. Uses N^2 evaluations over square.

GRASS.calc_mu — Method

Get mu on disk from provided (x,y) position: (x,y) = position on projected disk TODO: relies on small angle approx

GRASS.calc_r2 — Method

Calculate length of distance vector projected on face of star

GRASS.calc_rms — Method

TODO: this docstring calc rms about the observed mean

GRASS.calc_rvs_from_ccf — Method

GRASS.clean_input — Method

GRASS.doppler_factor — Method

GRASS.extrapolate_bisector — Method

GRASS.extrapolate_width — Method

GRASS.find_nearest_ax — Method

Find the nearest axis to a given point on a grid

GRASS.make_grid — Method

Make iterator representing uniform NxN grid

GRASS.mu_map — Method

Calculate mu for each position on a grid

GRASS.norm_term — Method

GRASS.patch_velocity_los — Method

patchvelocitylos(x, y; rstar, pole) Compute line of sight velocity of a patch of stellar surface given by x,y (assumed in [-1,1]). Optional params:

rstar = 1: in R_sol (affects return velocity, but not x,y)

pole = (0,1,0): unit vector with stellar rotation axis

Depends on:

GRASS.quad_limb_darkening — Method

Quadratic limb darkening law. Takes μ = cos(heliocentric angle) and LD parameters, u1 and u2.

GRASS.relative_bisector_wavelengths — Method

GRASS.rotation_period — Method

Calculate stellar rotation period in days:

sin_lat = sine of stellar latitude

A, B, C = Diff. rotation equation parameters (in deg/day)

GRASS.searchsortednearest — Method

Credit: traktofon @ https://discourse.julialang.org/t/findnearest-function/4143/4

GRASS.synthesize_spectra — Method

synthesize_spectra(spec, disk; seed_rng=false, verbose=true, top=NaN)

Synthesize spectra given parameters in spec and disk instances.

Arguments

spec::SpecParams: SpecParams instance

disk::DiskParams: DiskParams instance

GRASS.uniqueidx — Method

https://stackoverflow.com/questions/50899973/indices-of-unique-elements-of-vector-in-julia

GRASS.wav2vel — Method

GRASS.width_thermal — Method

Width of line due to thermal broadening