Index

GRASS.DiskParams

GRASS.SolarData

GRASS.SpecParams

GRASS.calc_ccf

GRASS.calc_rms

GRASS.calc_rvs_from_ccf

GRASS.patch_velocity_los

GRASS.rotation_period

GRASS.synthesize_spectra

GRASS.DiskParams — Method

DiskParams(; N=132, Nt=50, pole=(0.0, 1.0, 0.0))

Construct a DiskParams composite type instance.

Arguments

N::Integer=132: Length of N*N spatial grid

Nt::Integer=50: Number of 15-second snapshots

pole::Tuple{Float64, Float64, Float64}=(0.0, 1.0, 0.0): Unit vector specificying rotation axis direction. Default is equator-on.

GRASS.SolarData — Method

SolarData(; dir=soldir, relative=true, ...)

Construct a SpecParams composite type instance.

Arguments

dir::String=soldir: Directory containing the pre-processed input data. Default directory is set in src/config.jl

relative::Bool=true: Set whether wavelengths are on absolute scale or expressed relative to rest wavelength.

GRASS.SpecParams — Method

SpecParams(; lines=[], depths=[], variability=[], resolution=7e5)

Construct a SpecParams composite type instance. If variability is not specified, all lines are variable by default.

Arguments

lines::AbstractArray{Float64,1}=[]: List of line centers (in angstroms)

depths::AbstractArray{Float64,1}=[]: List of line depths

variability::AbstractArray{Bool,1}=[]: Array of booleans controlling whether corresponding line has variability.

resolution::Float64=7e8: Spectral resolution of spectrum

GRASS.calc_ccf — Method

calc_ccf(lambdas, intensities, lines, depths, resolution; normalize=true)

Compute the cross correlation function from a spectrum (lambdas and intensities) with a mask computed from a line list (lines).

Arguments

lambdas::AbstractArray{Float64,1}: List of wavelengths in spectrum.

intensities::AbstractArray{Float64,1}: List of intensities in spectrum.

lines::AbstractArray{Float64,1}: List of line centers for use in CCF mask.

depths::AbstractArray{Float64,1}: List of line depths for use as weights in CCF mask.

resolution::Float64: Spectral resolution of spcectrum.

GRASS.calc_rms — Method

calc_rms(A)

Calculate the RMS of given data about the observed mean.

Arguments

A::AbstractArray{Float64,1}: 1D array containing data for RMS calculation

GRASS.calc_rvs_from_ccf — Method

calc_rvs_from_ccf(v_grid, ccf)

Calculate apparent radial velocity from a CCF and velocity grid.

Arguments

v_grid::AbstractArray{Float64,1}: List of velocities returned by calc_ccf.

ccf::AbstractArray{Float64,1}: CCF values returned by calc_ccf.

GRASS.patch_velocity_los — Method

patch_velocity_los(x, y; rstar, pole)

Compute line of sight velocity of a patch of stellar surface given by x,y (assumed in [-1,1]). Return value is in (Rsol/day)/speed of light (i.e., dimensionless like z = v/c)

Arguments

rstar::Float64=1.0: in R_sol (affects return velocity, but not x,y)

pole = (0,1,0): unit vector for stellar rotation axis (default is equator-on)

GRASS.rotation_period — Method

rotation_period(sin_lat; A=14.713, B=2.396, C=1.787)

Calculate stellar rotation period in days at given sine of latitude.

Arguments

sin_lat::Float64: Sine of stellar latitude.

A::Float64=14.713: Coefficient from Snodgrass & Ulrich (1990)

B::Float64=2.396: Coefficient from Snodgrass & Ulrich (1990)

C::Float64=1.787: Coefficient from Snodgrass & Ulrich (1990)

GRASS.synthesize_spectra — Method

synthesize_spectra(spec, disk; seed_rng=false, verbose=true, top=NaN)

Synthesize spectra given parameters in spec and disk instances.

Arguments

spec::SpecParams: SpecParams instance

disk::DiskParams: DiskParams instance